+++ src/assets/svgs/index.tsx
